PS2: Refactor gamepad handling into its own file
It's independent enough that we can move it out of `main` and into
its own file.
